requirements.txt
numpy==1.19.0
tensorflow==2.0.0
keras==2.3.1
pandas==1.1.5
seaborn==0.11.0
scikit-learn==0.24.2
nltk==3.5.0
pyspellchecker==0.6.2
gensim==3.8.0
textblob==0.15.3
smart_open==2.0.0
xgboost==1.4.0
imbalanced-learn==0.8.0
scikit-optimize==0.8.1
